mail

+++ app/Http/Controllers/UsuarioController.php
@@ -15,6 +15,7 @@ class UsuarioController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', App\Models\Cliente::class);
         $gerentes =  Gerente::with('imagenMo')->get();
         $clientes = Cliente::with('imagenMo')->get();
         $usuarios = $gerentes->concat($clientes);
@@ -23,11 +24,13 @@ class UsuarioController extends Controller
 
     public function create()
     {
+        $this->authorize('create', App\Models\Cliente::class);
         return view("usuarios.agregar");
     }
 
     public function store(StoreUsuarioRequest $request)
     {
+        $this->authorize('create', App\Models\Cliente::class);
         if ($request->input('rol') == 'Gerente') {
             $usuario = new Gerente();
             $tipo =  Gerente::class;
@@ -55,13 +58,14 @@ class UsuarioController extends Controller
         return redirect(route('usuarios.index'));
     }
 
-    public function show(Usuario $usuario)
+    public function show($usuario)
     {
         //
     }
 
     public function edit($tipoUsuario, $id)
     {
+
         if ($tipoUsuario == 'Gerente') {
             $usuario = Gerente::where('id', $id)->first();
             if (is_null($usuario)) return view('errors.404');
@@ -69,7 +73,7 @@ class UsuarioController extends Controller
             $usuario = Cliente::where('id', $id)->first();
             if (is_null($usuario)) return view('errors.404');
         }
-
+        $this->authorize('update', $usuario);
         return view('usuarios.editar', compact('usuario', 'tipoUsuario'));
     }
 
@@ -82,11 +86,12 @@ class UsuarioController extends Controller
             $usuario = Cliente::where('id', $id)->first();
             $tipo =  Cliente::class;
         }
+        $this->authorize('update', $usuario);
 
         $usuario->fill($request->all());
-        $usuario->contraseña = Hash::make($request->input('contraseña'));
+     
         $usuario->save();
-        
+
         if ($request->hasFile('imagen')) {
             $imagen = $request->file('imagen');
             $ruta = 'imagenes/';
@@ -116,6 +121,7 @@ class UsuarioController extends Controller
         } else {
             $usuario = Cliente::where('id', $id)->first();
         }
+        $this->authorize('delete', $usuario);
 
         $usuario->delete();
         return redirect(route('usuarios.index'));
